Finalized initial logger implementation with heading support
